Changed chkfilelen to use long long. Fixed a few bugs in the twopass FFT routines.  Still don't work corectly.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@380 5fac236c-a0d8-0310-9745-ef116914eda1

+++ src/chkio.c
@@ -83,7 +83,7 @@ int chkfileseek(FILE * stream, long offset, size_t size, int whence)
 }
 
 
-unsigned long chkfilelen(FILE *file, size_t size)
+long long chkfilelen(FILE *file, size_t size)
 {
   int filenum, rt;
 #ifdef USE_PIOFS
@@ -107,7 +107,7 @@ unsigned long chkfilelen(FILE *file, size_t size)
     exit(-1);
   }
 #endif
-  return (unsigned long) (buf.st_size / size);
+  return (long long) (buf.st_size / size);
 }
 
 int read_int(FILE *infile)
